Filter block tools in fill wrapper
